+++ models/Fields.php
@@ -6,11 +6,11 @@ use Yii;
 use wdmg\base\models\ActiveRecordML;
 use wdmg\forms\models\Forms;
 use yii\behaviors\BlameableBehavior;
-use yii\behaviors\SluggableBehavior;
 use yii\behaviors\TimestampBehavior;
 use yii\db\ActiveRecord;
 use yii\db\Expression;
 use yii\helpers\ArrayHelper;
+use wdmg\base\behaviors\SluggableBehavior;
 
 /**
  * This is the model class for table "{{%forms_fields}}".
@@ -73,32 +73,14 @@ class Fields extends ActiveRecordML
      */
     public function behaviors()
     {
-        return [
-            'timestamp' => [
-                'class' => TimestampBehavior::class,
-                'attributes' => [
-                    ActiveRecordML::EVENT_BEFORE_INSERT => 'created_at',
-                    ActiveRecordML::EVENT_BEFORE_UPDATE => 'updated_at',
-                ],
-                'value' => new Expression('NOW()'),
-            ],
-            'blameable' => [
-                'class' => BlameableBehavior::class,
-                'createdByAttribute' => 'created_by',
-                'updatedByAttribute' => 'updated_by',
-            ],
-            'sluggable' =>  [
-                'class' => SluggableBehavior::class,
-                'attribute' => ['label'],
-                'slugAttribute' => 'name',
-                'ensureUnique' => true,
-                'skipOnEmpty' => true,
-                'immutable' => true,
-                'value' => function ($event) {
-                    return mb_substr($this->label, 0, 32);
-                }
-            ],
+        $behaviors = parent::behaviors();
+        $behaviors['sluggable'] = [
+            'class' => SluggableBehavior::class,
+            'attribute' => 'label',
+            'slugAttribute' => 'name',
+            'replacement' => '_'
         ];
+        return $behaviors;
     }
 
     /**
@@ -111,6 +93,8 @@ class Fields extends ActiveRecordML
             [['form_id', 'type', 'sort_order'], 'integer'],
             [['params'], 'string'],
             [['label', 'name'], 'string', 'max' => 64],
+            ['name', 'match', 'pattern' => '/^[A-za-z]/', 'message' => Yii::t('app/modules/forms','The attribute must begin with a letter.')],
+            ['name', 'match', 'pattern' => '/^[A-Za-z0-9\_]+$/', 'message' => Yii::t('app/modules/forms','It allowed only Latin alphabet, numbers and «_» character.')],
             [['status', 'is_required'], 'boolean'],
             [['placeholder'], 'string', 'max' => 124],
             [['description'], 'string', 'max' => 255],